# docs/swagger.yaml
definitions:
  handler.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handler.SignInResponse:
    properties:
      token:
        type: string
    type: object
  handler.SignUpRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      password_confirmation:
        type: string
    required:
    - email
    - password
    - password_confirmation
    type: object
  handler.SignUpResponse:
    properties:
      user:
        $ref: '#/definitions/handler.UserResponse'
    type: object
  handler.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
    type: object
info:
  contact: {}
paths:
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: SignIn payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SignInResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sign In
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: SignUp payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.SignUpResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sign Up
      tags:
      - auth
  /user/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve user information using the user ID
      parameters:
      - description: User ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful Response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid User ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get User by ID
      tags:
      - user
swagger: "2.0"